Testing a log for user check-in
The last two functions are just tests to see if the I could keep a record of the check ins. I am trying to use it with the membership class,  but I am not sure if it will work the way that I am thinking, take a look over and see if there is a way to implement both classes rather than creating a new class.